refactor: flutter에서 계산로직을 utils/로 분리
