+++ app/lib/src/pages/project_page/project_page.dart
@@ -12,7 +12,7 @@ import 'package:talk_pilot/src/pages/project_page/widgets/script_part_page.dart'
 import 'package:talk_pilot/src/pages/project_page/widgets/project_info_card.dart';
 import 'package:talk_pilot/src/pages/project_page/widgets/script_upload_button.dart';
 import 'package:talk_pilot/src/pages/project_page/widgets/editable/editable_text_editor.dart';
-import 'package:talk_pilot/src/services/project/estimated_time_service.dart';
+import 'package:talk_pilot/src/utils/project/estimated_time_service.dart';
 
 class ProjectPage extends StatefulWidget {
   final String projectId;
