+++ app/lib/src/utils/project/script_progress_service.dart
