+++ app/lib/src/utils/text_extract/txt_extract_service.dart
